fix(blocks): change order of block placement actions to fix protection (#177)
Currently, protection plugin checks are called on block place before we check if a custom block is being placed. We should only cancel the block place event if we are sure the block belongs to our plugin.

This also removes the check if canceled and instead passes it to the event handler.